feat(api): Add authorizeUrl API
Similar to `authorize` API, but returns URL instead of open the URL immediately.
